Implementação dos métodos da Classe curso + criação da Main para testes

+++ src/Professor.kt
@@ -1,5 +1,5 @@
 abstract class Professor( var nome : String,
                           var sobrenome : String,
-                          var tempoDeCasa : Short,
+                          var tempoDeCasa : Int,
                           val codigoDoProfessor : Int) {
 }
