+++ src/ProfessorAdjunto.kt
@@ -1,6 +1,6 @@
-class ProfessorAdjunto ( nome: String ,
+ class ProfessorAdjunto ( nome: String ,
                          sobrenome: String ,
-                         tempoDeCasa: Short ,
+                         tempoDeCasa: Int ,
                          codigoDoProfessor: Int ,
                          var qtdHorasMonitoria : Double) : Professor(nome , sobrenome , tempoDeCasa , codigoDoProfessor)  {
 }
