+++ src/ProfessorTitular.kt
@@ -1,6 +1,6 @@
 class ProfessorTitular(nome: String ,
                        sobrenome: String ,
-                       tempoDeCasa: Short ,
+                       tempoDeCasa: Int ,
                        codigoDoProfessor: Int ,
                        var especialidade: String) : Professor(nome , sobrenome , tempoDeCasa , codigoDoProfessor) {
 }
